python/requirements.txt
aiohttp>=3.8.0
aiofiles>=23.0.0
python-telegram-bot>=20.0
slack-sdk>=3.23.0
psutil>=5.9.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0